refactor evaluation and train hooks into checkpoint logging and metric hooks
