+++ .gitignore
@@ -4,8 +4,8 @@
 
 # Playing around stuff
 **play**
-**logs**
-**log**
+logs/
+log/
 
 # Pictures
 **.png
